Better compartmentalization for prebuilt components
