+++ blocks/src/prebuilt/index.js
@@ -1,0 +1,5 @@
+import text from './text'
+
+export default {
+    text
+}
